Split out carbon emissions from cap
So we will always spit out carbon emissions even if we are nog running
a carbon cap.
