+++ gridpath/project/operations/operational_types/gen_commit_bin.py
@@ -2299,6 +2299,14 @@ def fuel_rule(mod, g):
         return None
 
 
+def carbon_emissions_rule(mod, g, tmp):
+    if g in mod.GEN_COMMIT_BIN_FUEL_PRJS:
+        return mod.GenCommitBin_Fuel_Burn_MMBTU[g, tmp] \
+            * mod.co2_intensity_tons_per_mmbtu[mod.gen_commit_bin_fuel[g]]
+    else:
+        return 0
+
+
 def variable_om_cost_rule(mod, g, tmp):
     """
     Variable O&M cost has two components which are additive:
