+++ gridpath/project/operations/operational_types/gen_simple.py
@@ -544,6 +544,16 @@ def fuel_rule(mod, g):
         return None
 
 
+def carbon_emissions_rule(mod, g, tmp):
+    if g in mod.GEN_SIMPLE_FUEL_PRJS:
+        return mod.gen_simple_fuel_burn_slope_mmbtu_per_mwh[
+                   g, mod.period[tmp], 0] \
+            * mod.Power_Provision_MW[g, tmp] \
+            * mod.co2_intensity_tons_per_mmbtu[mod.gen_simple_fuel[g]]
+    else:
+        return 0
+
+
 def startup_cost_rule(mod, g, tmp):
     """
     Since there is no commitment, there is no concept of starting up.
