+++ gridpath/project/operations/operational_types/stor.py
@@ -619,6 +619,12 @@ def fuel_rule(mod, g):
     return None
 
 
+def carbon_emissions_rule(mod, g, tmp):
+    """
+    """
+    return 0
+
+
 def variable_om_cost_rule(mod, g, tmp):
     """
     Variable O&M costs are applied only to the storage discharge, i.e. when the
