+++ tests/project/operations/test_carbon_emissions.py
@@ -61,102 +61,6 @@ class TestCarbonEmissions(unittest.TestCase):
                               stage=""
                               )
 
-    def test_load_model_data(self):
-        """
-        Test that data are loaded with no errors
-        :return:
-        """
-        add_components_and_load_data(prereq_modules=IMPORTED_PREREQ_MODULES,
-                                     module_to_test=MODULE_BEING_TESTED,
-                                     test_data_dir=TEST_DATA_DIRECTORY,
-                                     subproblem="",
-                                     stage=""
-                                     )
-
-    def test_data_loaded_correctly(self):
-        """
-        Test components initialized with data as expected
-        :return:
-        """
-        m, data = add_components_and_load_data(
-            prereq_modules=IMPORTED_PREREQ_MODULES,
-            module_to_test=MODULE_BEING_TESTED,
-            test_data_dir=TEST_DATA_DIRECTORY,
-            subproblem="",
-            stage=""
-        )
-        instance = m.create_instance(data)
-
-        # Set: CRBN_PRJS
-        expected_carbonaceous_projects = sorted(
-            ["Gas_CCGT", "Coal", "Gas_CT", "Gas_CCGT_New",
-             "Gas_CCGT_New_Binary", "Gas_CT_New",
-             "Gas_CCGT_z2", "Coal_z2", "Gas_CT_z2", "Disp_Binary_Commit",
-             "Disp_Cont_Commit", "Disp_No_Commit", "Clunky_Old_Gen",
-             "Clunky_Old_Gen2"])
-        actual_carbonaceous_projects = \
-            sorted([p for p in instance.CRBN_PRJS])
-        self.assertListEqual(expected_carbonaceous_projects,
-                             actual_carbonaceous_projects)
-
-        # Param: carbon_cap_zone
-        expected_cc_zone_by_prj = OrderedDict(sorted({
-            "Gas_CCGT": "Carbon_Cap_Zone1",
-            "Coal": "Carbon_Cap_Zone1",
-            "Gas_CT": "Carbon_Cap_Zone1",
-            "Gas_CCGT_New": "Carbon_Cap_Zone1",
-            "Gas_CCGT_New_Binary": "Carbon_Cap_Zone1",
-            "Gas_CT_New": "Carbon_Cap_Zone1",
-            "Gas_CCGT_z2": "Carbon_Cap_Zone2",
-            "Coal_z2": "Carbon_Cap_Zone2",
-            "Gas_CT_z2": "Carbon_Cap_Zone2",
-            "Disp_Binary_Commit": "Carbon_Cap_Zone1",
-            "Disp_Cont_Commit": "Carbon_Cap_Zone1",
-            "Disp_No_Commit": "Carbon_Cap_Zone1",
-            "Clunky_Old_Gen": "Carbon_Cap_Zone1",
-            "Clunky_Old_Gen2": "Carbon_Cap_Zone1",
-            }.items())
-        )
-        actual_cc_zone_by_prj = OrderedDict(sorted({
-            p: instance.carbon_cap_zone[p] for p in
-            instance.CRBN_PRJS}.items()
-                                                    )
-                                             )
-        self.assertDictEqual(expected_cc_zone_by_prj, actual_cc_zone_by_prj)
-
-        # Set: CRBN_PRJS_BY_CARBON_CAP_ZONE
-        expected_prj_by_zone = OrderedDict(sorted({
-            "Carbon_Cap_Zone1": sorted([
-                "Gas_CCGT", "Coal", "Gas_CT", "Gas_CCGT_New",
-                "Gas_CCGT_New_Binary", "Gas_CT_New",
-                "Disp_Binary_Commit", "Disp_Cont_Commit", "Disp_No_Commit",
-                "Clunky_Old_Gen", "Clunky_Old_Gen2"
-            ]),
-            "Carbon_Cap_Zone2": sorted(["Gas_CCGT_z2", "Coal_z2", "Gas_CT_z2"])
-                                                  }.items()
-                                                  )
-                                           )
-        actual_prj_by_zone = OrderedDict(sorted({
-            z: sorted([p for p in
-                       instance.CRBN_PRJS_BY_CARBON_CAP_ZONE[z]
-                ])
-            for z in instance.CARBON_CAP_ZONES
-                                                }.items()
-                                                )
-                                         )
-        self.assertDictEqual(expected_prj_by_zone, actual_prj_by_zone)
-
-        # Set: CRBN_PRJ_OPR_TMPS
-        expected_carb_prj_op_tmp = sorted(
-            get_project_operational_timepoints(expected_carbonaceous_projects)
-        )
-
-        actual_carb_prj_op_tmp = sorted([
-            (prj, tmp) for (prj, tmp)
-            in instance.CRBN_PRJ_OPR_TMPS
-        ])
-        self.assertListEqual(expected_carb_prj_op_tmp, actual_carb_prj_op_tmp)
-
 
 if __name__ == "__main__":
     unittest.main()
